1.Add Canbus detail message decode function(Cell voltage, Ntc voltage). 2.Add Module DVP & Module DTP & ID protect parameter. 3.Add Max Current & Peak current, Rate Voltage parameter. 4.Add BMU number of Modulr parameter. 5.App bms add each scu cell voltage, ntc voltage, soc. soh......information.

+++ Lib/SMP/smp_can_fifo.h
@@ -24,8 +24,8 @@ typedef struct{
 typedef struct{
 	smp_can_package_t	*buffer_addr;	             /* FIFO queue buffer start address */
 	uint16_t	buffer_size;	                     /* FIFO queue buffer size */
-	uint32_t	in;				                         /* FIFO queue buffer push in address */
-	uint32_t	out;			                        /* FIFO queue buffer pop out address */
+	uint16_t	in;				                         /* FIFO queue buffer push in address */
+	uint16_t	out;			                        /* FIFO queue buffer pop out address */
 }smp_can_fifo_t;
 
 /* Exported constants --------------------------------------------------------*/
